requirements.txt
lxml==4.9.4
python-docx==1.1.0
typing_extensions==4.9.0
